drivers: ieee802154: separate common methods from dw1000 driver
